requirements.txt
agate==1.7.1
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
attrs==25.3.0
azure-core==1.34.0
azure-identity==1.23.0
babel==2.17.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==14.0
cryptography==45.0.3
daff==1.4.2
dagster==1.10.18
dagster-dbt==0.26.18
dagster-pipes==1.10.18
dagster_shared==1.10.18
dbt-adapters==1.15.3
dbt-common==1.25.0
dbt-core==1.8.9
dbt-extractor==0.5.1
dbt-fabric==1.8.9
dbt-protos==1.0.317
dbt-semantic-interfaces==0.5.1
dbt-sqlserver==1.8.7
deepdiff==7.0.1
docstring_parser==0.16
filelock==3.18.0
fsspec==2025.5.1
greenlet==3.2.2
grpcio==1.71.0
grpcio-health-checking==1.62.3
humanfriendly==10.0
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
leather==0.4.0
Logbook==1.5.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.14
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.6
ordered-set==4.1.0
orjson==3.10.18
packaging==25.0
pandas==2.2.3
parsedatetime==2.6
pathspec==0.11.2
polars==1.30.0
protobuf==5.29.5
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.1.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
shellingham==1.5.4
six==1.17.0
snowplow-tracker==1.1.0
SQLAlchemy==2.0.41
sqlglot==26.24.0
sqlglotrs==0.5.0
sqlparse==0.5.3
structlog==25.3.0
tabulate==0.9.0
text-unidecode==1.3
tomli==2.2.1
tomlkit==0.13.2
toposort==1.10
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
universal_pathlib==0.2.6
urllib3==1.26.20
watchdog==5.0.3
zipp==3.22.0
